--- BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/IServiceLayer/IAuthAccounts.java
@@ -1,6 +1,0 @@
-package com.bauschbakeries.bauschbakeries_backend.IServiceLayer;
-
-public interface IAuthAccounts {
-
-
-}
