--- BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/IServiceLayer/IPermissions.java
@@ -1,4 +1,0 @@
-package com.bauschbakeries.bauschbakeries_backend.IServiceLayer;
-
-public interface IPermissions {
-}
