+++ BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/interfaces/IAuthAccounts.java
@@ -1,0 +1,6 @@
+package com.bauschbakeries.bauschbakeries_backend.interfaces;
+
+public interface IAuthAccounts {
+
+
+}
