+++ BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/interfaces/IPermissions.java
@@ -1,0 +1,4 @@
+package com.bauschbakeries.bauschbakeries_backend.interfaces;
+
+public interface IPermissions {
+}
