+++ BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/services/AuthAccountsService.java
@@ -1,16 +1,17 @@
-package com.bauschbakeries.bauschbakeries_backend.IServiceLayer;
+package com.bauschbakeries.bauschbakeries_backend.services;
 
+import com.bauschbakeries.bauschbakeries_backend.interfaces.IAuthAccounts;
 import com.bauschbakeries.bauschbakeries_backend.repositories.OAuthAccountsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class IAuthAccountsService implements IAuthAccounts {
+public class AuthAccountsService implements IAuthAccounts {
 
     @Autowired
     private final OAuthAccountsRepository AuthAccountsRepo;
 
-    public IAuthAccountsService(OAuthAccountsRepository AuthAccountsRepo) {
+    public AuthAccountsService(OAuthAccountsRepository AuthAccountsRepo) {
         this.AuthAccountsRepo = AuthAccountsRepo;
     }
