+++ BauschBakeries_Backend/src/main/java/com/bauschbakeries/bauschbakeries_backend/services/PermissionsService.java
@@ -1,16 +1,17 @@
-package com.bauschbakeries.bauschbakeries_backend.IServiceLayer;
+package com.bauschbakeries.bauschbakeries_backend.services;
 
+import com.bauschbakeries.bauschbakeries_backend.interfaces.IPermissions;
 import com.bauschbakeries.bauschbakeries_backend.repositories.PermissionsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class IPermissionsService implements IPermissions {
+public class PermissionsService implements IPermissions {
 
     @Autowired
     private final PermissionsRepository permissionsRepo;
 
-    public IPermissionsService(PermissionsRepository permissionsRepo) {
+    public PermissionsService(PermissionsRepository permissionsRepo) {
         this.permissionsRepo = permissionsRepo;
     }
 }
